VLDAP
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@61844 05bdf26c-840f-0410-9ced-eb539d925f36

+++ webs/vldap-web/docroot/WEB-INF/src/com/liferay/vldap/util/PortletPropsValues.java
@@ -28,7 +28,7 @@ public class PortletPropsValues {
 	public static final String BIND_SASL_HOSTNAME = GetterUtil.getString(
 		PortletProps.get("bind.sasl.hostname"));
 
-	public static final long SEARCH_MAX_SIZE = GetterUtil.getLong(
+	public static final int SEARCH_MAX_SIZE = GetterUtil.getInteger(
 		PortletProps.get("search.max.size"));
 
 	public static final int SEARCH_MAX_TIME = GetterUtil.getInteger(
